Fix free-space

+++ configuration.nix
@@ -198,21 +198,27 @@
          #
          # nix-collect-garbage --delete-old
 
-         docker ps --all --quiet | xargs --no-run-if-empty docker stop \
-         && docker ps --all --quiet | xargs --no-run-if-empty docker rm \
-         && docker images --quiet | xargs --no-run-if-empty docker rmi --force \
-         && docker container prune --force \
-         && docker image prune --force \
-         && docker network prune --force
+         if command -v docker &> /dev/null
+         then
+           docker ps --all --quiet | xargs --no-run-if-empty docker stop \
+           && docker ps --all --quiet | xargs --no-run-if-empty docker rm \
+           && docker images --quiet | xargs --no-run-if-empty docker rmi --force \
+           && docker container prune --force \
+           && docker image prune --force \
+           && docker network prune --force
+         fi                 
 
-         podman pod prune --force
-         podman ps --all --quiet | xargs --no-run-if-empty podman rm --force \
-         && podman images --quiet | xargs --no-run-if-empty podman rmi --force \
-         && podman container prune --force \
-         && podman images --quiet | podman image prune --force \
-         && podman network ls --quiet | xargs --no-run-if-empty podman network rm \
-         && podman pod list --quiet | xargs --no-run-if-empty podman pod rm --force
-
+         if command -v docker &> /dev/null
+         then
+           podman pod prune --force
+    
+           podman ps --all --quiet | xargs --no-run-if-empty podman rm --force \
+           && podman images --quiet | xargs --no-run-if-empty podman rmi --force \
+           && podman container prune --force \
+           && podman images --quiet | podman image prune --force \
+           && podman network ls --quiet | xargs --no-run-if-empty podman network rm \
+           && podman pod list --quiet | xargs --no-run-if-empty podman pod rm --force
+         fi
          nix profile remove '.*' \
          && find ~ \( -iname '*.iso' -o -iname '*.qcow2*' -o -iname '*.img' -o -iname 'result' \) -exec rm -fv -- {} + 2> /dev/null | tr ' ' '\n' \
          && sudo rm -fr "$HOME"/.cache "$HOME"/.local \
